% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{featurizeChat}
\alias{featurizeChat}
\title{Featurize Chats}
\usage{
featurizeChat(chat_data, feature_name = NA, featurization_function, ...)
}
\arguments{
\item{chat_data}{A list of chat data generated from the parseChats function.}

\item{feature_name}{Optional name for the feature to be added to chat_data.  If not provided, the feature will be named after the featurization function.}

\item{featurization_function}{The function to be used to featurize each chat message.}

\item{...}{Additional arguments to be passed to the featurization_function}
}
\description{
This function generates features to describe each message in chats, using any featurization_function the user passes, so long as the function returns a vector the same length as the number of messages in each chat.
}
\examples{
featurizeChat()
}
\keyword{featurize}
