% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getAlterVars}
\alias{getAlterVars}
\title{Get Alter Variables}
\usage{
getAlterVars(survey_data, var_names)
}
\arguments{
\item{survey_data}{A dataframe of survey responses, including a column named ego_code that contains the receipt code the respondent entered after completing their chat.}

\item{var_names}{A vector of variable names, that are present in colnames(survey_data), for which each participant's alter's values are desired.}
}
\description{
This function retrieves the values of each participant's partner's (specified) variables, and appends them to the survey_data dataframe as new columns named as paste0("alter_", gsub("ego_", "", var_name)).  Note that this function is written for dyadic chats only.
}
\examples{
getAlterVars()
}
\keyword{alter}
