% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parseChat}
\alias{parseChat}
\title{Parse Chats}
\usage{
parseChat(file)
}
\arguments{
\item{file}{Path to csv file.}
}
\description{
This function parses a ReChat csv export into an R list.
}
\examples{
parseChat()
}
\keyword{parse}
