% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{printChat}
\alias{printChat}
\title{Print Chat}
\usage{
printChat(
  chat,
  width = 500,
  height = 500,
  file = "chat_export",
  format = "pdf"
)
}
\arguments{
\item{chat}{The chat to print, for example chat_data[\link{1}]}

\item{width}{Width of output file in pixels.}

\item{height}{Height of output file in pixels.}

\item{file}{File path to save export.}

\item{format}{File format for export.  Currently "pdf" and "png" are supported.}
}
\description{
This function exports a selected chat as a readable PDF file.
}
\examples{
printChat()
}
\keyword{alter}
