% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{summarizeChat}
\alias{summarizeChat}
\title{Summarize Chat Features}
\usage{
summarizeChat(
  survey_data,
  chat_data,
  chat_feature_name,
  summary_feature_name = NA,
  summary_function,
  ...
)
}
\arguments{
\item{survey_data}{A dataframe of survey responses, including a column named ego_code that contains the receipt code the respondent entered after completing their chat.}

\item{chat_data}{A list of chat data generated from the parseChat function.}

\item{chat_feature_name}{The name of the chat feature (added to chat_data using featurizeChat) to summarize.}

\item{summary_feature_name}{Optional name for the summary feature to be added to survey_data.  If not provided, a descriptive name will be automatically generated as paste0("ego_",enexpr(summary_function),"_",chat_feature_name)}

\item{summary_function}{A function, such as mean(), to use to summarize the message-level features of a chat to the participant-level.}

\item{...}{Additional arguments to be passed to the summary_function.}
}
\description{
This function takes chat features generated by featurizeChat, summarizes them to the participant level, and adds this summary information to the survey_data dataframe.
}
\examples{
summarizeChat()
}
\keyword{summarize}
